refactor: rename helper to utils
