+++ utils/map_helper.go
@@ -1,4 +1,4 @@
-package helper
+package utils
 
 import (
 	"errors"
